Redis caching working

+++ backend-services/ticket-service/src/main/java/com/supportai/ticket/dto/TicketDTO.java
@@ -4,6 +4,7 @@ import com.supportai.ticket.domain.TicketPriority;
 import com.supportai.ticket.domain.TicketStatus;
 import lombok.Data;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Data
 public class TicketDTO {
@@ -23,5 +24,5 @@ public class TicketDTO {
  
     private boolean aiProcessed;
     private String aiResponse;
-    private Double aiConfidence;
+    private BigDecimal aiConfidence;
 }
